pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[project]
name = "slida"
authors = [
    {name = "Robert Huselius", email = "robert@huseli.us"},
]
description = ""
license = {file = "LICENSE"}
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pillow",
    "platformdirs",
    "PySide6",
    "pyyaml",
    "klaatu-python@https://github.com/Eboreg/klaatu-python/archive/refs/heads/master.zip",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "flake8",
    "ipdb",
    "ipython",
    "isort",
    "pylint",
]

[project.scripts]
slida = "slida.application:main"

[tool.setuptools.dynamic]
version = {attr = "slida.__version__"}

[tool.pylint.main]
disable = [
    "broad-exception-caught",
    "import-outside-toplevel",
    "invalid-name",
    "invalid-str-returned",
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
    "too-few-public-methods",
    "too-many-ancestors",
    "too-many-arguments",
    "too-many-branches",
    "too-many-instance-attributes",
    "too-many-locals",
    "too-many-positional-arguments",
    "too-many-public-methods",
    "too-many-return-statements",
    "too-many-statements",
    "unused-argument",
]
ignore = [".venv", "node_modules"]
ignore-patterns = ["^\\.#", ".*\\.pyi", "apps\\.py"]
max-line-length = 119
extension-pkg-allow-list = ["PySide6.QtWidgets", "PySide6"]

[tool.isort]
combine_as_imports = true
default_section = "THIRDPARTY"
include_trailing_comma = true
line_length = 79
multi_line_output = 3
lines_after_imports = 2
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
skip = [".venv", "node_modules"]

[tool.mypy]
follow_imports = "skip"
show_error_codes = true
ignore_missing_imports = true
files = "."
check_untyped_defs = true

[tool.pyright]
exclude = ["**/__pycache__", "build", "**/.*", ".venv"]
typeCheckingMode = "basic"
ignore = [".venv"]
